Created Navigation between pages/functionality tweeks
Added the hamberger navigation button to all pages and linked all pages to eachother. Added Spencer's home-scripts page to 3 of the pages, and added code to the java.js. Tweeked css. pages along with html pages to be able to function properly with the hamberger. Almsot all pages had links such as style or index pages that were not unique and were copying overeachother, so I had to create new pages for muscan and rename just about all of the pages.